test(livedoc): run signature crypto tests under node env, not jsdom
jsdom runs test code in an isolated vm context, so ArrayBuffer/Uint8Array
instances built there fail Node's native WebCrypto instanceof-ArrayBuffer
check on importKey, even though crypto.subtle itself is Node's real
SubtleCrypto. This made verifySignature() swallow a spurious TypeError
into `false` and fail the P-256 round-trip test - a jsdom/Node realm
artifact, not a bug in liveDocSignature.ts (single-realm browsers/webviews
aren't affected). Split the crypto describe block out of
LiveDocInsert.test.ts into its own file tagged
`// @vitest-environment node`.
